requirements.txt
Flask==2.2.3
grpcio==1.51.3
protobuf==4.22.3
